feat(bundle): split published plugin id into its own module registering local + external
Moves PlugwrightPlugin (the io.github.drownek.plugwright entry point)
out of plugwright-local into a new plugwright-bundle module that
applies the core engine and registers both built-in modes. Mirrors
plugwright-local's old jar-merging trick, now pulling in api, core,
local and external classes since none of them publish standalone
coordinates.

plugwright-local goes back to being just a mode module - no publish
plugin, no plugin-id registration - matching plugwright-external's
shape. Published plugin id and artifact coordinates are unchanged, so
existing consumer build scripts keep working.
