+++ gradle-plugin/plugwright-bundle/src/main/kotlin/me/drownek/plugwright/PlugwrightPlugin.kt
@@ -1,5 +1,6 @@
 package me.drownek.plugwright
 
+import me.drownek.plugwright.external.ExternalMode
 import me.drownek.plugwright.local.LocalMode
 import org.gradle.api.Plugin
 import org.gradle.api.Project
@@ -7,14 +8,15 @@ import org.gradle.api.Project
 /**
  * Entry point for the `io.github.drownek.plugwright` id.
  *
- * Applies the mode-agnostic engine and registers the built-in modes — just `local` for
- * now. A dedicated bundle module can take over this role once a second built-in mode
- * exists to combine with it.
+ * Applies the mode-agnostic engine and registers both built-in modes: `local` and `external`.
+ * A third-party mode registers itself the same way, from its own plugin or from the build
+ * script directly, via `plugwright.registerMode(...)`.
  */
 class PlugwrightPlugin : Plugin<Project> {
     override fun apply(project: Project) {
         project.pluginManager.apply(PlugwrightCorePlugin::class.java)
         val extension = project.extensions.getByType(PlugwrightExtension::class.java)
         extension.registerMode(LocalMode)
+        extension.registerMode(ExternalMode)
     }
 }
